+++ rbac-distributed-service-sdk/src/main/java/com/dwarfeng/rbacds/sdk/bean/entity/WebInputUser.java
@@ -1,9 +1,11 @@
 package com.dwarfeng.rbacds.sdk.bean.entity;
 
 import com.alibaba.fastjson.annotation.JSONField;
+import com.dwarfeng.rbacds.sdk.util.Constraints;
 import com.dwarfeng.rbacds.stack.bean.entity.User;
 import com.dwarfeng.subgrade.sdk.bean.key.WebInputStringIdKey;
 import com.dwarfeng.subgrade.stack.bean.Bean;
+import org.hibernate.validator.constraints.Length;
 
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
@@ -17,7 +19,7 @@ import java.util.Objects;
  */
 public class WebInputUser implements Bean {
 
-    private static final long serialVersionUID = 3139169583343730301L;
+    private static final long serialVersionUID = 2681324024929702832L;
 
     public static User toStackBean(WebInputUser webInputUser) {
         if (Objects.isNull(webInputUser)) {
@@ -35,6 +37,7 @@ public class WebInputUser implements Bean {
     private WebInputStringIdKey key;
 
     @JSONField(name = "remark")
+    @Length(max = Constraints.LENGTH_REMARK)
     private String remark;
 
     public WebInputUser() {