+++ rbac-distributed-service-sdk/src/main/java/com/dwarfeng/rbacds/sdk/util/Constraints.java
@@ -28,6 +28,13 @@ public final class Constraints {
      */
     public static final int LENGTH_NAME = 20;
 
+    /**
+     * 用户的长度。
+     *
+     * @since 1.8.0
+     */
+    public static final int LENGTH_USER = 50;
+
     private Constraints() {
         throw new IllegalStateException("禁止实例化");
     }